+++ src/django_elasticsearch_dsl_drf/filter_backends/__init__.py
@@ -14,7 +14,10 @@ from .ordering import (
     OrderingFilterBackend,
 )
 from .search import SearchFilterBackend
-from .suggester import SuggesterFilterBackend
+from .suggester import (
+    SuggesterFilterBackend,
+    FunctionalSuggesterFilterBackend,
+)
 from .highlight import HighlightBackend
 
 __title__ = 'django_elasticsearch_dsl_drf.filter_backends'